Выделил функции в отдельный файл
